+++ buti_app/threads/__init__.py
@@ -1,9 +1,9 @@
 from .serial_thread import SerialThread
 from .sdk_camera_thread import SDKCameraThread
-from .micromanager_camera_thread import MicroManagerCameraThread
+from .mmcore_camera_thread import MMCoreCameraThread
 
 __all__ = [
     "SerialThread",
     "SDKCameraThread",
-    "MicroManagerCameraThread",
+    "MMCoreCameraThread",
 ]